Add support for #lang-less assembly files (code from Shu-Hung You)
Thanks very much to Shu-Hung You for this contribution via the racket-users
mailing list.

This adjusts the definition of read-syntax so that it produces a module that
Racket can recognize (using #'module from br/quicklang; strip-bindings, as
previously used, removed the binding of module), and then in main constructs
a suitable namespace in which to eval the resulting module.